feat(project):实现项目管理功能

+++ biz/internal/service/project/project.go
@@ -5,7 +5,9 @@ import (
 	"github.com/cloudwego/hertz/pkg/app"
 	"github.com/cloudwego/hertz/pkg/common/utils"
 	"github.com/cloudwego/hertz/pkg/protocol/consts"
+	"personal-page-be/biz/internal/constant"
 	"personal-page-be/biz/internal/domain"
+	"personal-page-be/biz/internal/service/error_type"
 	"strconv"
 )
 
@@ -22,34 +24,52 @@ func (s *ProjectService) AddProject(ctx context.Context, c *app.RequestContext) 
 	err = s.Repo.SaveProject(&project)
 	if err != nil {
 		c.JSON(consts.StatusOK, utils.H{
-			"code": 5001,
+			"code": error_type.ErrInternal.Code,
 			"msg":  err.Error(),
 		})
 		return
 	}
 	c.JSON(consts.StatusOK, utils.H{
 		"code": 0,
-		"msg":  "添加成功",
+		"data": project.ID,
 	})
 }
 
 func (s *ProjectService) RemoveProject(ctx context.Context, c *app.RequestContext) {
+	err0, _ := s.GlobalService.CheckLogin(c, true)
+	if err0 != nil {
+		c.JSON(consts.StatusOK, utils.H{
+			"code": err0.Code,
+			"msg":  err0.Message,
+		})
+		return
+	}
+
 	projectID := c.Param("id")
 	projectIDInt, err := strconv.ParseUint(projectID, 10, 64)
 	if err != nil {
 		c.JSON(consts.StatusOK, utils.H{
-			"code": 2001,
-			"msg":  "参数错误",
+			"code": error_type.ErrBadRequest.Code,
+			"msg":  error_type.ErrBadRequest.Message,
 		})
 		return
 	}
 
 	project, err := s.Repo.GetProject(uint(projectIDInt))
+
 	if err != nil {
 		c.JSON(consts.StatusOK, utils.H{
 			"code": 5001,
 			"msg":  err.Error(),
 		})
+		return
+	}
+	if project.ID == 0 {
+		c.JSON(consts.StatusOK, utils.H{
+			"code": error_type.ErrNotFound.Code,
+			"msg":  error_type.ErrNotFound.Message,
+		})
+		return
 	}
 
 	err = s.Repo.RemoveProject(project.ID)
@@ -67,18 +87,47 @@ func (s *ProjectService) RemoveProject(ctx context.Context, c *app.RequestContex
 	return
 }
 
-func (s *ProjectService) GetPages(ctx context.Context, c *app.RequestContext) {
-
+func (s *ProjectService) GetNum(ctx context.Context, c *app.RequestContext) {
+	count, err := s.Repo.GetProjectsNum()
+	if err != nil {
+		c.JSON(consts.StatusOK, utils.H{
+			"code": 5001,
+			"msg":  err.Error(),
+		})
+	}
 	c.JSON(consts.StatusOK, utils.H{
 		"code": 0,
-		"msg":  "获取成功",
+		"data": count,
 	})
 }
 
 func (s *ProjectService) GetProjects(ctx context.Context, c *app.RequestContext) {
+	start := c.DefaultQuery("start", "0")
+	startInt, err := strconv.Atoi(start)
+	if err != nil {
+		c.JSON(consts.StatusOK, utils.H{
+			"code": 4001,
+			"msg":  err.Error(),
+		})
+	}
+
+	end := c.DefaultQuery("end", strconv.Itoa(startInt+10))
+	endInt, err := strconv.Atoi(end)
+	if err != nil {
+		c.JSON(consts.StatusOK, utils.H{
+			"code": 4001,
+			"msg":  err.Error(),
+		})
+	}
+
+	if endInt-startInt > constant.MaxProjectsNum {
+		endInt = startInt + constant.MaxProjectsNum + 1
+	}
+
+	projects, err := s.Repo.GetProjects(startInt, endInt)
 
 	c.JSON(consts.StatusOK, utils.H{
 		"code": 0,
-		"msg":  "获取成功",
+		"data": projects,
 	})
 }
